+++ packages/webdriverio/test/esmTest.mjs
@@ -1,15 +1,17 @@
 import defaultExport from '../dist/index.mjs';
+import { AxeBuilder } from '../dist/index.mjs';
 import assert from 'assert';
 import * as webdriverio from 'webdriverio';
 import { fileURLToPath, pathToFileURL } from 'url';
 import { join } from 'path';
+import { fixturesPath } from 'axe-test-fixtures';
 
-const exportIsFunction = typeof defaultExport === 'function';
-assert(exportIsFunction, 'export is not a function');
+assert(typeof defaultExport === 'function', 'default export is not a function');
+assert(typeof AxeBuilder === 'function', 'named export is not a function')
+assert(defaultExport === AxeBuilder, 'default and named export are not the same');
 
 async function integrationTest() {
-  let path = fileURLToPath(new URL('.', import.meta.url));
-  path = join(path, '../fixtures/external/index.html');
+  const path = join(fixturesPath, 'index.html');
 
   const options = {
     automationProtocol: 'devtools',