+++ packages/webdriverjs/test/esmTest.mjs
@@ -1,5 +1,8 @@
+// ensure compatibility of ESM format
 import defaultExport from '../dist/index.mjs';
+import { AxeBuilder } from '../dist/index.mjs';
 import assert from 'assert';
 
-const exportIsFunction = typeof defaultExport === 'function';
-assert(exportIsFunction, 'export is not a function');
+assert(typeof defaultExport === 'function', 'default export is not a function');
+assert(typeof AxeBuilder === 'function', 'named export is not a function')
+assert(defaultExport === AxeBuilder, 'default and named export are not the same');